+++ src/apps/toolbar/modules/Device/infra.tsx
@@ -1,6 +1,6 @@
 import { DeviceTM } from 'seelen-core';
 
-import { Item } from '../item/infra';
+import { Item } from '../item/infra/infra';
 
 interface Props {
   module: DeviceTM;